refactor: move the table to widgets
